+++ frontend/app/src/main/java/kr/ac/tukorea/whereareu/data/repository/setting/SettingRepositoryImpl.kt
@@ -4,6 +4,7 @@ import kr.ac.tukorea.whereareu.data.api.SettingService
 import kr.ac.tukorea.whereareu.data.model.ResponseBody
 import kr.ac.tukorea.whereareu.data.model.setting.ModifyUserInfoRequest
 import kr.ac.tukorea.whereareu.data.model.setting.ModifyUserInfoResponse
+import kr.ac.tukorea.whereareu.domain.login.userinfo.GetUserInfoResult
 import kr.ac.tukorea.whereareu.util.network.NetworkResult
 import kr.ac.tukorea.whereareu.util.network.handleApi
 import javax.inject.Inject
@@ -13,5 +14,9 @@ class SettingRepositoryImpl @Inject constructor(
 ) : SettingRepository{
     override suspend fun sendModifyUserInfo(request: ModifyUserInfoRequest): NetworkResult<ModifyUserInfoResponse> {
         return handleApi({ api.postModifyUserInfo(request) }) { response: ResponseBody<ModifyUserInfoResponse> -> response.result }
+    }
+
+    override suspend fun getUserInfo(dementiaKey: String): NetworkResult<GetUserInfoResult> {
+        return handleApi({api.getUserInfo(dementiaKey)}) {response: ResponseBody<GetUserInfoResult> -> response.result}
     }
 }
