--- frontend/app/src/main/java/kr/ac/tukorea/whereareu/data/repository/userInfo/UserInfoRepository.kt
@@ -1,9 +1,0 @@
-package kr.ac.tukorea.whereareu.data.repository.userInfo
-
-import kr.ac.tukorea.whereareu.data.model.setting.GetUserInfoResponse
-import kr.ac.tukorea.whereareu.domain.login.userinfo.GetUserInfoResult
-import kr.ac.tukorea.whereareu.util.network.NetworkResult
-
-interface UserInfoRepository {
-    suspend fun getUserInfo(dementiaKey: String): NetworkResult<GetUserInfoResult>
-}
