--- frontend/app/src/main/java/kr/ac/tukorea/whereareu/data/repository/userInfo/UserInfoRepositoryImpl.kt
@@ -1,17 +1,0 @@
-package kr.ac.tukorea.whereareu.data.repository.userInfo
-
-import kr.ac.tukorea.whereareu.data.api.UserInfoService
-import kr.ac.tukorea.whereareu.data.model.ResponseBody
-import kr.ac.tukorea.whereareu.data.model.setting.GetUserInfoResponse
-import kr.ac.tukorea.whereareu.domain.login.userinfo.GetUserInfoResult
-import kr.ac.tukorea.whereareu.util.network.NetworkResult
-import kr.ac.tukorea.whereareu.util.network.handleApi
-import javax.inject.Inject
-
-class UserInfoRepositoryImpl @Inject constructor(
-    private val api: UserInfoService
-) : UserInfoRepository{
-    override suspend fun getUserInfo(dementiaKey: String): NetworkResult<GetUserInfoResult> {
-        return handleApi({api.getUserInfo(dementiaKey)}) {response: ResponseBody<GetUserInfoResult> -> response.result}
-    }
-}
