+++ frontend/app/src/main/java/kr/ac/tukorea/whereareu/di/NetworkModule.kt
@@ -10,7 +10,6 @@ import kr.ac.tukorea.whereareu.WhereAreUApplication
 import kr.ac.tukorea.whereareu.data.api.dementia.DementiaHomeService
 import kr.ac.tukorea.whereareu.data.api.LoginService
 import kr.ac.tukorea.whereareu.data.api.SettingService
-import kr.ac.tukorea.whereareu.data.api.UserInfoService
 import kr.ac.tukorea.whereareu.data.api.nok.NokHomeService
 import kr.ac.tukorea.whereareu.util.location.LocationService
 import okhttp3.Interceptor
