+++ frontend/app/src/main/java/kr/ac/tukorea/whereareu/presentation/nok/setting/NokSettingFragment.kt
@@ -7,7 +7,6 @@ import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
 import kr.ac.tukorea.whereareu.R
 import kr.ac.tukorea.whereareu.databinding.FragmentNokSettingBinding
-import kr.ac.tukorea.whereareu.presentation.UserInfoViewModel
 import kr.ac.tukorea.whereareu.presentation.base.BaseFragment
 import kr.ac.tukorea.whereareu.presentation.nok.home.NokHomeViewModel
 import kr.ac.tukorea.whereareu.util.extension.repeatOnStarted
