+++ frontend/app/src/main/java/kr/ac/tukorea/whereareu/util/getUserInfo/GetUserInfoService.kt
@@ -4,7 +4,6 @@ import android.app.Service
 import android.content.Intent
 import android.os.IBinder
 import dagger.hilt.android.AndroidEntryPoint
-import kr.ac.tukorea.whereareu.data.repository.userInfo.UserInfoRepositoryImpl
 import javax.inject.Inject
 
 @AndroidEntryPoint
